+++ wolfcrypt/src/dsa.c
@@ -347,7 +347,7 @@ int wc_DsaSign(const byte* digest, byte* out, DsaKey* key, WC_RNG* rng)
     int    ret, sz;
     byte   buffer[DSA_HALF_SIZE];
 
-    sz = min(sizeof(buffer), mp_unsigned_bin_size(&key->q));
+    sz = min((int)sizeof(buffer), mp_unsigned_bin_size(&key->q));
 
     /* generate k */
     ret = wc_RNG_GenerateBlock(rng, buffer, sz);
